requirements.txt
Django>=3.1.6,< 3.2.0
djangorestframework>=3.12.2,<3.13.0
flake8>=3.6.0,<3.7.0
psycopg2>=2.8.6,<2.9.0